+++ src/main/kotlin/com/rankweis/uppercut/karate/psi/formatter/GherkinCodeStyleSettingsProvider.java
@@ -2,13 +2,21 @@ package com.rankweis.uppercut.karate.psi.formatter;
 
 import com.intellij.application.options.CodeStyleAbstractConfigurable;
 import com.intellij.application.options.CodeStyleAbstractPanel;
+import com.intellij.lang.Language;
 import com.intellij.psi.codeStyle.CodeStyleConfigurable;
 import com.intellij.psi.codeStyle.CodeStyleSettings;
 import com.intellij.psi.codeStyle.CodeStyleSettingsProvider;
 import com.rankweis.uppercut.karate.MyBundle;
+import com.rankweis.uppercut.karate.psi.KarateLanguage;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public final class GherkinCodeStyleSettingsProvider extends CodeStyleSettingsProvider {
+
+  @Override
+  public @Nullable Language getLanguage() {
+    return KarateLanguage.INSTANCE;
+  }
   
   @NotNull
   @Override